Create pw.txt
the app needs a plaintext file called pw.txt in the same directory to work. This is the database where the passwords are kept. This file is NEVER to be uploaded or shared ANYWHERE, as it is in plaintext. Don't put it on a network drive or something like that either.